worked on u8 vector -> ir, and edited hex and binary -> u8 vector functions
