+++ tools/data-conversion/src/u8vector.rs
@@ -1,10 +1,14 @@
-pub fn binary_to_u8_vec(binary: &str) -> Result<Vec<u8>, String> {
-    let mut padded_binary: String = binary.to_string();
+use crate::ir::IntermediateRepresentation;
+use num_bigint::{BigUint, BigInt};
 
-    // If the binary string length is not a multiple of 8, pad it with leading zeros
+pub fn binary_to_u8_vec(binary: &str) -> Result<Vec<u8>, String> {
+    let mut padded_binary = binary.to_string();
+
+    // If the binary string length is not a multiple of 8, pad with the most significant bit
     let padding = 8 - (padded_binary.len() % 8);
     if padding != 8 {
-        padded_binary = "0".repeat(padding) + &padded_binary;
+        let msb = &padded_binary[0..1]; // Get the most significant bit
+        padded_binary = msb.repeat(padding) + &padded_binary; // Pad with the MSB
     }
 
     let mut vec = Vec::new();
@@ -20,39 +24,54 @@ pub fn binary_to_u8_vec(binary: &str) -> Result<Vec<u8>, String> {
     Ok(vec)
 }
 
+
 pub fn hex_to_u8_vec(hex: &str) -> Result<Vec<u8>, String> {
     let mut padded_hex = hex.to_string();
 
+    // If the hex string length is not a multiple of 2, pad with the most significant nibble
     let padding = 2 - (padded_hex.len() % 2);
     if padding != 2 {
-        padded_hex = "0".repeat(padding) + &padded_hex;
+        let msb = &padded_hex[0..1]; // Get the most significant nibble
+        padded_hex = msb.repeat(padding) + &padded_hex; // Pad with the MSB
     }
 
     let mut vec = Vec::new();
 
     for i in (0..padded_hex.len()).step_by(2) {
-        let byte_str = &padded_hex[i..i + 8];
-        match u8::from_str_radix(byte_str, 2){
+        let byte_str = &padded_hex[i..i + 2]; // Fixed to extract two hex digits
+        match u8::from_str_radix(byte_str, 16) {
             Ok(byte) => vec.push(byte),
-            Err(_) => return Err(String::from("Invalid binary string")),
+            Err(_) => return Err(String::from("Invalid hex string")),
         }
     }
     Ok(vec)
 }
 
 
-pub fn decimal_to_u8_vec(decimal: &str) -> Result<Vec<u8>, String> {
-    let mut vec = Vec::new();
 
-    // Iterate over each character in the decimal string
-    for c in decimal.chars() {
-        // Check if the character is a digit
-        if let Some(digit) = c.to_digit(10) {
-            // Convert the digit (u32) to a u8 and push it to the vector
-            vec.push(digit as u8);
-        } else {
-            return Err(format!("Invalid character '{}' in decimal string", c));
+pub fn u8_to_ir(vector: Result<Vec<u8>, String>, exponent: i64) -> IntermediateRepresentation {
+    match vector {
+        Ok(vec) => {
+            // Check if the MSB of the first byte is 1
+            let is_negative = (vec[0] & 0b10000000) != 0; 
+
+            let mantissa = if is_negative {
+                // Convert the Vec<u8> to a two's complement BigInt, then get its absolute value
+                let bigint = BigInt::from_signed_bytes_be(&vec);
+                bigint.magnitude().clone() // absolute value 
+            } else {
+                BigUint::from_bytes_be(&vec)
+            };
+
+            IntermediateRepresentation {
+                sign: is_negative,
+                mantissa,
+                exponent,
+            }
+        }
+        Err(e) => {
+            // Handle the error case, for example by panicking or returning a default value.
+            panic!("Failed to convert: {}", e);
         }
     }
-    Ok(vec)
 }
